enfore `cdbg_block_groups` contract

+++ products/cdbg/models/staging/stg__census_data_blocks.sql
@@ -1,6 +1,6 @@
 SELECT
     geoid20 AS bctcb2020,
-    borocode AS borough_code,
+    borocode::integer AS borough_code,
     geogname AS borough_name,
     "pop1.1"::numeric AS total_population   -- noqa: RF01
 FROM {{ source("recipe_sources", "dcp_censusdata_blocks") }}
